wayland: Avoid passing wl_fixed as ints
The current code in calculateDragPosition took an integer argument that
wasn't a integer value, but a casted wl_fixed.

It did the right thing with it, but it's confusing to read.

Instead convert all wl_fixed results as early as possible and pass
qreal's around to be clearer.

+++ src/plugins/platforms/wayland/qwaylanddatadevice_p.h
@@ -85,7 +85,7 @@ private Q_SLOTS:
 
 private:
 #if QT_CONFIG(draganddrop)
-    QPoint calculateDragPosition(int x, int y, QWindow *wnd) const;
+    QPoint calculateDragPosition(qreal x, qreal y, QWindow *wnd) const;
 #endif
     void sendResponse(Qt::DropActions supportedActions, const QPlatformDragQtResponse &response);
